requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.1
cron_descriptor==2.0.6
Django==5.2.10
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
eventlet==0.40.4
greenlet==3.3.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.52
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
